+++ packages/style-kit/tsup.config.js
@@ -1,10 +1,24 @@
 import { defineConfig } from "tsup";
 
-export default defineConfig({
-  clean: true,
-  dts: true,
-  entryPoints: ["src/**/*"],
-  format: ["esm"],
-  outDir: "dist",
-  sourcemap: true,
-});
+export default defineConfig([
+  {
+    clean: true,
+    dts: true,
+    entryPoints: ["src/**/*", "!src/**/*.test.ts"],
+    format: ["esm"],
+    outDir: "dist",
+    sourcemap: true,
+  },
+  {
+    banner: {
+      js: "#!/usr/bin/env node",
+    },
+    clean: false,
+    dts: false,
+    entryPoints: ["bin/index.ts"],
+    format: ["esm"],
+    noExternal: ["commander"],
+    outDir: "dist/bin",
+    sourcemap: true,
+  },
+]);